project refactor
move files, rename project, add tests projects
