--- TagsterWebAPI/Tagster.Domain/Models/Tag.cs
@@ -1,9 +1,0 @@
-﻿namespace Tagster.Domain.Models
-{
-    public class Tag
-    {
-        public int Id { get; set; }
-        public string TagName { get; set; }
-        public int ProfileId { get; set; }
-    }
-}
