--- TagsterWebAPI/TagsterWebAPI/Tagster.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace TagsterWebAPI
-{
-    public class Tagster
-    {
-        public string Terms { get; set; }
-    }
-}
