+++ TagsterWebAPI/libraries/Tagster.DataAccess/DbContexts/ITagsterDbContext.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Tagster.DataAccess.Entities;
+
+namespace Tagster.DataAccess.DBContexts
+{
+    public interface ITagsterDbContext
+    {
+        DbSet<Tag> Tags { get; set; }
+        DbSet<Profile> Profiles { get; set; }
+    }
+}
