+++ TagsterWebAPI/libraries/Tagster.DataAccess/DbContexts/TagsterDbContext.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Tagster.DataAccess.Entities;
+
+namespace Tagster.DataAccess.DBContexts
+{
+    public class TagsterDbContext : DbContext, ITagsterDbContext
+    {
+        public DbSet<Tag> Tags { get; set; }
+        public DbSet<Profile> Profiles { get; set; }
+
+        public TagsterDbContext(DbContextOptions options) : base(options)
+        {
+                
+        }
+    }
+}
