+++ TagsterWebAPI/libraries/Tagster.DataAccess/Entities/Tag.cs
@@ -1,0 +1,9 @@
+﻿namespace Tagster.DataAccess.Entities
+{
+    public class Tag
+    {
+        public int Id { get; set; }
+        public string TagName { get; set; }
+        public int ProfileId { get; set; }
+    }
+}
