+++ TagsterWebAPI/src/Tagster.Application/Extensions/Extension.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Tagster.Application.Extensions
+{
+    public static class Extension
+    {
+        public static IServiceCollection AddApplication(this IServiceCollection services) 
+            => services;
+    }
+}
