+++ TagsterWebAPI/src/Tagster.Application/Interfaces/ITagsterDbContext.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Tagster.DataAccess.Entities;
+
+namespace Tagster.Application.Interfaces
+{
+    public interface ITagsterDbContext
+    {
+        DbSet<Tag> Tags { get; set; }
+        DbSet<Profile> Profiles { get; set; }
+    }
+}
