+++ TagsterWebAPI/src/Tagster.Application/Tagster.Application.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.8" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\libraries\Tagster.DataAccess\Tagster.DataAccess.csproj" />
+    <ProjectReference Include="..\Tagster.Domain\Tagster.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
