+++ TagsterWebAPI/src/Tagster.Infrastructure/Tagster.Infrastructure.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\libraries\Tagster.DataAccess\Tagster.DataAccess.csproj" />
+    <ProjectReference Include="..\Tagster.Application\Tagster.Application.csproj" />
+    <ProjectReference Include="..\Tagster.Domain\Tagster.Domain.csproj" />
+  </ItemGroup>
+
+
+</Project>
