+++ TagsterWebAPI/src/TagsterWebAPI/Program.cs
@@ -1,0 +1,18 @@
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
+
+namespace TagsterWebAPI
+{
+    public class Program
+    {
+        public static void Main(string[] args) 
+            => CreateHostBuilder(args).Build().Run();
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
+    }
+}
