+++ TagsterWebAPI/src/TagsterWebAPI/Tagster.cs
@@ -1,0 +1,7 @@
+namespace TagsterWebAPI
+{
+    public class Tagster
+    {
+        public string Terms { get; set; }
+    }
+}
