+++ TagsterWebAPI/tests/Tagster.Application.UnitTests/UnitTest1.cs
@@ -1,0 +1,12 @@
+using Xunit;
+
+namespace Tagster.Application.UnitTests
+{
+    public class UnitTest1
+    {
+        //dummy test
+        [Fact]
+        public void Test1() 
+            => Assert.True(true);
+    }
+}
