+++ TagsterWebAPI/tests/Tagster.DataAccess.UnitTests/UnitTest1.cs
@@ -1,0 +1,13 @@
+using System;
+using Xunit;
+
+namespace Tagster.DataAccess.UnitTests
+{
+    public class UnitTest1
+    {
+        //dummy test
+        [Fact]
+        public void Test1()
+            => Assert.True(true);
+    }
+}
